+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - 全屏 Loading 现在默认不再锁定屏幕滚动。如果需要的话，可添加 `lock` 修饰符
 - Table 删除属性 fixedColumnCount, customCriteria, customBackgroundColors
 - Table 的 selectionchange、cellmouseenter、cellmouseleave、cellclick 事件更名为 selection-change、cell-mouse-enter、cell-mouse-leave、cell-click。
+- Pagination 的 currentchange、sizechange 事件更名为 current-change、size-change
 
 ### 1.0.0-rc.7